pyproject.toml
[project]
name = "nlp-project-grp-6"
version = "1.0.0"
description = "NLP Project from Group 6"
authors = [
    {name = "Your Name",email = "you@example.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<3.15"
dependencies = [
    "jupyterlab (>=4.4.10,<5.0.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "scipy (>=1.16.3,<2.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "accelerate (>=1.11.0,<2.0.0)",
    "datasets (>=4.3.0,<5.0.0)",
    "evaluate (>=0.4.6,<0.5.0)",
    "plotly (>=6.3.1,<7.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "jupyterlab-execute-time (>=3.2.0,<4.0.0)",
    "rouge-score (>=0.1.2,<0.2.0)",
    "nltk (>=3.9.2,<4.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "torch (>=2.5.1,<3.0.0)",
    "torchvision (>=0.20.1,<0.21.0)",
    "torchaudio (>=2.5.1,<3.0.0)",
    "transformers (<4.41)",
    "sentencepiece (>=0.2.1,<0.3.0)",
    "youtube-transcript-api (>=1.2.3,<2.0.0)",
    "vadersentiment (>=3.3.2,<4.0.0)",
    "ipywidgets (>=8.1.8,<9.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "scrapetube (>=2.6.0,<3.0.0)",
    "yt-dlp[curl-cffi,default] (>=2025.11.12,<2026.0.0)",
    "openai (>=2.8.1,<3.0.0)",
    "azure-identity (>=1.25.1,<2.0.0)",
    "azure-ai-projects (>=2.0.0b1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pytorch-gpu"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"


[tool.poetry.dependencies]
torch = {source = "pytorch-gpu"}
torchvision = {source = "pytorch-gpu"}
torchaudio = {source = "pytorch-gpu"}
